feature/LW-8.Display user actions. Refactoring of code
